Move props to TileControl.Properties.cs
